+++ hdp/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod cairo_runner;
-pub mod config;
 pub mod constant;
 pub mod hdp_run;
 pub mod preprocessor;
